Refactor firebase components
Move firebase components from main.ts to firebaseApp.ts, update imports and exports. Update docs.
