+++ docs/index.365cb23a.js
@@ -1611,5 +1611,5 @@ function(e){Ne="9.6.10",Z(new L("database",((e,{instanceIdentifier:t})=>Br(e.get
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-ie("firebase","9.6.10","app");const Vr=function(e=function(e="[DEFAULT]"){const t=Q.get(e);if(!t)throw te.create("no-app",{appName:e});return t}(),t){return ee(e,"database").getImmediate({identifier:t})}(function(e,t={}){if("object"!=typeof t){t={name:t}}const n=Object.assign({name:"[DEFAULT]",automaticDataCollectionEnabled:!1},t),i=n.name;if("string"!=typeof i||!i)throw te.create("bad-app-name",{appName:String(i)});const s=Q.get(i);if(s){if(S(e,s.options)&&S(n,s.config))return s;throw te.create("duplicate-app",{appName:i})}const r=new U(i);for(const e of X.values())r.addComponent(e);const o=new ne(e,n,r);return Q.set(i,o),o}({apiKey:"AIzaSyBSQ4EwRXTEF6Rs_3REQTWtcOoBu3HOKEQ",authDomain:"gritbook-social-media-app.firebaseapp.com",databaseURL:"https://gritbook-social-media-app-default-rtdb.europe-west1.firebasedatabase.app",projectId:"gritbook-social-media-app",storageBucket:"gritbook-social-media-app.appspot.com",messagingSenderId:"110931333305",appId:"1:110931333305:web:c540e52ab897acade50b68"}));class $r{constructor(e,t){this.username=e,this.password=t}getUsername(){return this.username}getPassword(){return this.password}}const Yr=document.querySelector("#travel"),Gr=document.querySelector("#sport"),Kr=document.querySelector("#gaming"),Qr=document.querySelector("#travel-chat"),Xr=document.querySelector("#sport-chat"),Jr=document.querySelector("#gaming-chat");let Zr=Yr,eo=Qr,to=!0,no=!1,io=!1;const so=e=>{"travel-forum"==e?(Zr=Yr,eo=Qr,Yr.style.display="flex",Gr.style.display="none",Kr.style.display="none",to=!0,no=!1,io=!1,vo()):"sport-forum"==e?(Zr=Gr,eo=Xr,Yr.style.display="none",Gr.style.display="flex",Kr.style.display="none",to=!1,no=!0,io=!1,vo()):"gaming-forum"==e&&(Zr=Kr,eo=Jr,Yr.style.display="none",Gr.style.display="none",Kr.style.display="flex",to=!1,no=!1,io=!0,vo())},ro=e=>{const t=document.querySelector(".user"),n=document.createElement("div"),i=document.createElement("div"),s=document.createElement("h5"),r=document.createElement("h5"),o=document.createElement("p"),a=document.createElement("button"),l=document.createElement("div");s.innerText=e.username,r.innerText=e.timestamp,o.innerText=e.message,a.innerText="X",a.className="delete-btn",n.className="user-div",i.className="mess-div",l.className="message-container",l.id=e.id,n.append(s),n.append(r),i.append(o),e.username==t.innerText&&e.userId==t.id&&(i.append(a),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Zr.id);const t=Rr(Vr,"/messages/"+Zr.id+"/"+e.id);console.log(t),Mr(t)}))),l.append(n),l.append(i),eo.prepend(l),Zr.append(eo),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Zr.id);const t=Rr(Vr,"/messages/"+Zr.id+"/"+e.id);console.log(t),Mr(t)}))};class oo{constructor(e,t,n,i,s){this.id=e,this.username=t,this.message=n,this.timestamp=i,this.userId=s}clearChat(){document.querySelector(`#${this.id}`).remove()}}const ao=Rr(Vr,"/messages/travel"),lo=Rr(Vr,"messages/sport"),ho=Rr(Vr,"messages/gaming"),co=Rr(Vr,"/users");let uo;const po=document.querySelectorAll(".forum-btn"),_o=document.querySelector(".user"),fo=document.querySelector("#message"),go=document.querySelector("#add-message-btn");let mo=[];const yo=()=>{!0===to?uo=ao:!0===no?uo=lo:!0===io&&(uo=ho)},vo=()=>{yo(),Wr(uo,(e=>{const t=e.val();console.log(t);for(const e of mo)e.clearChat();mo=[];for(const e in t)mo.push(new oo(e,t[e].username,t[e].message,t[e].timestamp,t[e].userId));console.log(mo);for(const e of mo)ro(e)}))},Co=()=>{let e;return Wr(co,(t=>{e=t.val(),console.log(e)})),e};Co(),new class{constructor(){this.signInForm=document.querySelector("#login-form"),this.signUpForm=document.querySelector("#signup-form"),this.username=document.querySelector("#username"),this.password=document.querySelector("#password"),document.querySelector("#signin-btn").addEventListener("click",this.validator.bind(this)),document.querySelector("#btn-signup-ui").addEventListener("click",this.hideUI.bind(this))}validator(e){if(e.preventDefault(),console.log("validator invoked"),console.log(this.username.value),console.log(this.password.value),this.username.value.length>=4&&this.password.value.length>=4){const e=Co();if(e){for(const t in e)if(this.username.value==e[t].username&&this.password.value==e[t].password){console.log("login success");const n=document.querySelector(".user");return n.textContent=e[t].username,n.setAttribute("id",t),void this.clearForm()}console.log("login failed"),this.clearForm()}}else console.log("username/password length is less than 4 characters"),this.clearForm()}hideUI(e){e.preventDefault(),this.clearForm(),this.signInForm.style.display="none",this.signUpForm.style.display="flex"}clearForm(){this.signInForm.reset()}},new class{constructor(){this.usernameInField=document.querySelector("#new_username"),this.passwordInField=document.querySelector("#new_password"),this.repeatPasswordInField=document.querySelector("#repeat_password"),this.signUpform=document.querySelector("#signup-form"),this.signInForm=document.querySelector("#login-form"),document.querySelector("#signup-btn").addEventListener("click",this.validateForm.bind(this)),document.querySelector("#btn-display-signin").addEventListener("click",this.hideUI.bind(this))}validateForm(e){e.preventDefault();const t=Co();let n=[];for(const e in t)n.push(t[e].username);this.usernameInField.value.length<4||this.passwordInField.value.length<4||this.repeatPasswordInField.value.length<4?alert("username/password length is less than 4 characters"):n.includes(this.usernameInField.value)?alert("username already exists"):this.passwordInField.value==this.repeatPasswordInField.value?(alert("signup success"),this.initNewUser(new $r(this.usernameInField.value,this.passwordInField.value))):alert("password does not match")}hideUI(e){e.preventDefault(),console.log("hideUI invoked"),this.signUpform.style.display="none",this.signInForm.style.display="flex"}initNewUser(e){console.log("initNewUser invoked");const t={username:e.getUsername(),password:e.getPassword()},n=Rr(Vr,"/users"),i={};i[Ar(n).key]=t,Or(n,i),this.clearForm()}clearForm(){this.signUpform.reset()}},so("travel-forum"),po.forEach((e=>{e.addEventListener("click",(e=>{const t=e.target.id;so(t)}))})),go.addEventListener("click",(e=>{e.preventDefault(),yo();const t=Date.now(),n=new Date(t),i=e=>e<10?"0"+e:e,s=i(n.getDate())+"-"+i(n.getMonth()+1)+"-"+n.getFullYear()+", "+i(n.getHours())+":"+i(n.getMinutes());if(""==fo.value)alert("Enter text to post!");else{const e={message:fo.value,username:_o.innerText,timestamp:s,userId:_o.id};fo.value="";const t={};t[Ar(uo).key]=e,Or(uo,t)}}));
-//# sourceMappingURL=index.20c0bba9.js.map
+ie("firebase","9.6.10","app");const Vr=document.querySelector("#travel"),$r=document.querySelector("#sport"),Yr=document.querySelector("#gaming"),Gr=document.querySelector("#travel-chat"),Kr=document.querySelector("#sport-chat"),Qr=document.querySelector("#gaming-chat");let Xr=Vr,Jr=Gr,Zr=!0,eo=!1,to=!1;const no=e=>{"travel-forum"==e?(Xr=Vr,Jr=Gr,Vr.style.display="flex",$r.style.display="none",Yr.style.display="none",Zr=!0,eo=!1,to=!1,_o()):"sport-forum"==e?(Xr=$r,Jr=Kr,Vr.style.display="none",$r.style.display="flex",Yr.style.display="none",Zr=!1,eo=!0,to=!1,_o()):"gaming-forum"==e&&(Xr=Yr,Jr=Qr,Vr.style.display="none",$r.style.display="none",Yr.style.display="flex",Zr=!1,eo=!1,to=!0,_o())},io=e=>{const t=document.querySelector(".user"),n=document.createElement("div"),i=document.createElement("div"),s=document.createElement("h5"),r=document.createElement("h5"),o=document.createElement("p"),a=document.createElement("button"),l=document.createElement("div");s.innerText=e.username,r.innerText=e.timestamp,o.innerText=e.message,a.innerText="X",a.className="delete-btn",n.className="user-div",i.className="mess-div",l.className="message-container",l.id=e.id,n.append(s),n.append(r),i.append(o),e.username==t.innerText&&e.userId==t.id&&(i.append(a),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Xr.id);const t=Rr(ro,"/messages/"+Xr.id+"/"+e.id);console.log(t),Mr(t)}))),l.append(n),l.append(i),Jr.prepend(l),Xr.append(Jr),a.addEventListener("click",(()=>{console.log("delete button clicked"),console.log(e.id),console.log(Xr.id);const t=Rr(ro,"/messages/"+Xr.id+"/"+e.id);console.log(t),Mr(t)}))};class so{constructor(e,t,n,i,s){this.id=e,this.username=t,this.message=n,this.timestamp=i,this.userId=s}clearChat(){document.querySelector(`#${this.id}`).remove()}}const ro=function(e=function(e="[DEFAULT]"){const t=Q.get(e);if(!t)throw te.create("no-app",{appName:e});return t}(),t){return ee(e,"database").getImmediate({identifier:t})}(function(e,t={}){if("object"!=typeof t){t={name:t}}const n=Object.assign({name:"[DEFAULT]",automaticDataCollectionEnabled:!1},t),i=n.name;if("string"!=typeof i||!i)throw te.create("bad-app-name",{appName:String(i)});const s=Q.get(i);if(s){if(S(e,s.options)&&S(n,s.config))return s;throw te.create("duplicate-app",{appName:i})}const r=new U(i);for(const e of X.values())r.addComponent(e);const o=new ne(e,n,r);return Q.set(i,o),o}({apiKey:"AIzaSyBSQ4EwRXTEF6Rs_3REQTWtcOoBu3HOKEQ",authDomain:"gritbook-social-media-app.firebaseapp.com",databaseURL:"https://gritbook-social-media-app-default-rtdb.europe-west1.firebasedatabase.app",projectId:"gritbook-social-media-app",storageBucket:"gritbook-social-media-app.appspot.com",messagingSenderId:"110931333305",appId:"1:110931333305:web:c540e52ab897acade50b68"})),oo=Rr(ro,"/messages/travel"),ao=Rr(ro,"messages/sport"),lo=Rr(ro,"messages/gaming"),ho=Rr(ro,"/users");let co,uo=[];const po=()=>{!0===Zr?co=oo:!0===eo?co=ao:!0===to&&(co=lo)},_o=()=>{po(),Wr(co,(e=>{const t=e.val();console.log(t);for(const e of uo)e.clearChat();uo=[];for(const e in t)uo.push(new so(e,t[e].username,t[e].message,t[e].timestamp,t[e].userId));console.log(uo);for(const e of uo)io(e)}))},fo=()=>{let e;return Wr(ho,(t=>{e=t.val(),console.log(e)})),e};fo();class go{constructor(e,t){this.username=e,this.password=t}getUsername(){return this.username}getPassword(){return this.password}}const mo=document.querySelectorAll(".forum-btn"),yo=document.querySelector(".user"),vo=document.querySelector("#message"),Co=document.querySelector("#add-message-btn");new class{constructor(){this.signInForm=document.querySelector("#login-form"),this.signUpForm=document.querySelector("#signup-form"),this.username=document.querySelector("#username"),this.password=document.querySelector("#password"),document.querySelector("#signin-btn").addEventListener("click",this.validator.bind(this)),document.querySelector("#btn-signup-ui").addEventListener("click",this.hideUI.bind(this))}validator(e){if(e.preventDefault(),console.log("validator invoked"),console.log(this.username.value),console.log(this.password.value),this.username.value.length>=4&&this.password.value.length>=4){const e=fo();if(e){for(const t in e)if(this.username.value==e[t].username&&this.password.value==e[t].password){console.log("login success");const n=document.querySelector(".user");return n.textContent=e[t].username,n.setAttribute("id",t),void this.clearForm()}console.log("login failed"),this.clearForm()}}else console.log("username/password length is less than 4 characters"),this.clearForm()}hideUI(e){e.preventDefault(),this.clearForm(),this.signInForm.style.display="none",this.signUpForm.style.display="flex"}clearForm(){this.signInForm.reset()}},new class{constructor(){this.usernameInField=document.querySelector("#new_username"),this.passwordInField=document.querySelector("#new_password"),this.repeatPasswordInField=document.querySelector("#repeat_password"),this.signUpform=document.querySelector("#signup-form"),this.signInForm=document.querySelector("#login-form"),document.querySelector("#signup-btn").addEventListener("click",this.validateForm.bind(this)),document.querySelector("#btn-display-signin").addEventListener("click",this.hideUI.bind(this))}validateForm(e){e.preventDefault();const t=fo();let n=[];for(const e in t)n.push(t[e].username);this.usernameInField.value.length<4||this.passwordInField.value.length<4||this.repeatPasswordInField.value.length<4?alert("username/password length is less than 4 characters"):n.includes(this.usernameInField.value)?alert("username already exists"):this.passwordInField.value==this.repeatPasswordInField.value?(alert("signup success"),this.initNewUser(new go(this.usernameInField.value,this.passwordInField.value))):alert("password does not match")}hideUI(e){e.preventDefault(),console.log("hideUI invoked"),this.signUpform.style.display="none",this.signInForm.style.display="flex"}initNewUser(e){console.log("initNewUser invoked");const t={username:e.getUsername(),password:e.getPassword()},n=Rr(ro,"/users"),i={};i[Ar(n).key]=t,Or(n,i),this.clearForm()}clearForm(){this.signUpform.reset()}},no("travel-forum"),mo.forEach((e=>{e.addEventListener("click",(e=>{const t=e.target.id;no(t)}))})),Co.addEventListener("click",(e=>{e.preventDefault(),po();const t=Date.now(),n=new Date(t),i=e=>e<10?"0"+e:e,s=i(n.getDate())+"-"+i(n.getMonth()+1)+"-"+n.getFullYear()+", "+i(n.getHours())+":"+i(n.getMinutes());if(""==vo.value)alert("Enter text to post!");else{const e={message:vo.value,username:yo.innerText,timestamp:s,userId:yo.id};vo.value="";const t={};t[Ar(co).key]=e,Or(co,t)}}));
+//# sourceMappingURL=index.365cb23a.js.map
